pyproject.toml
[project]
name = "nql-bot"
version = "0.1.0"
description = "Natural Query Language chatbot for movie queries using LangChain"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "google-generativeai>=0.8.5",
    "langchain-google-genai>=2.0.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "mypy>=1.18.1",
    "pytest>=8.4.2",
]
